Проводит небольшой рефакторинг по разделению функций на моковые и общие, используемые в проекте
