+++ src/utils/common.js
@@ -1,0 +1,23 @@
+import dayjs from 'dayjs';
+import { MIN_IN_HOUR } from '../const.js';
+
+// Функции применяемые в проекте
+
+const getRuntimeInHours = (runtime) => {
+  const hours = Math.trunc(runtime / MIN_IN_HOUR);
+  if (hours === 0) {
+    return `${runtime % MIN_IN_HOUR}m`;
+  }
+  return `${Math.trunc(runtime / MIN_IN_HOUR)}h ${runtime % MIN_IN_HOUR}m`;
+};
+
+const humanizeYear = (date) => dayjs(date).format('YYYY');
+const humanizeCommentDate = (date) => dayjs(date).format('YYYY/MM/DD HH:mm');
+const humanizeReleaseDate = (date) => dayjs(date).format('DD MMMM YYYY');
+
+export {
+  getRuntimeInHours,
+  humanizeCommentDate,
+  humanizeReleaseDate,
+  humanizeYear
+};
