+++ src/utils/mocks.js
@@ -1,6 +1,3 @@
-import dayjs from 'dayjs';
-import { MIN_IN_HOUR } from './const.js';
-
 // Функции для получения моковых данных
 
 // Функция из интернета по генерации случайного числа из диапазона
@@ -39,28 +36,10 @@ const getArrayFromRandomElements = (arr) => {
   return Array.from(new Set(randomLengthArray));
 };
 
-// Функции применяемые в проекте
-
-const getRuntimeInHours = (runtime) => {
-  const hours = Math.trunc(runtime / MIN_IN_HOUR);
-  if (hours === 0) {
-    return `${runtime % MIN_IN_HOUR}m`;
-  }
-  return `${Math.trunc(runtime / MIN_IN_HOUR)}h ${runtime % MIN_IN_HOUR}m`;
-};
-
-const humanizeYear = (date) => dayjs(date).format('YYYY');
-const humanizeCommentDate = (date) => dayjs(date).format('YYYY/MM/DD HH:mm');
-const humanizeReleaseDate = (date) => dayjs(date).format('DD MMMM YYYY');
-
 export {
   getRandomInteger,
   getRandomFloat,
   generateRandomElement,
   getPoster,
-  getRuntimeInHours,
   getArrayFromRandomElements,
-  humanizeCommentDate,
-  humanizeReleaseDate,
-  humanizeYear
 };
